chore: refactor openUrlUtils
