+++ apps/desktop/src-electron/preload.ts
@@ -2,17 +2,11 @@
 /* eslint-disable @typescript-eslint/no-unused-vars,@typescript-eslint/require-await */
 import { ipcRenderer } from 'electron';
 
+import type { IPrefType } from '@onekeyhq/shared/types/desktop';
+
 import { ipcMessageKeys } from './config';
 
 import type { IUpdateSettings } from './libs/store';
-
-export type IPrefType =
-  | 'camera'
-  | 'bluetooth'
-  | 'location'
-  | 'notification'
-  | 'locationService'
-  | 'localNetwork';
 
 export type IDesktopAppState = 'active' | 'background' | 'blur';
