+++ packages/shared/types/desktop.ts
@@ -1,0 +1,7 @@
+export type IPrefType =
+  | 'camera'
+  | 'bluetooth'
+  | 'location'
+  | 'notification'
+  | 'locationService'
+  | 'localNetwork';
